Change: Move new safety system classes to separate namespace
