+++ source/OpenBVE/Simulation/TrainManager/Train/SafetySystems/PilotLamp.cs
@@ -2,53 +2,50 @@
 using OpenBveApi.Trains;
 using SoundManager;
 
-namespace OpenBve
+namespace OpenBve.SafetySystems
 {
-	public partial class TrainManager
+	internal class PilotLamp
 	{
-		internal class PilotLamp
+		/// <summary>Played once when all doors are closed</summary>
+		internal CarSound OnSound;
+		/// <summary>Played once when the first door opens</summary>
+		internal CarSound OffSound;
+		/// <summary>Holds the reference to the base car</summary>
+		private readonly AbstractCar baseCar;
+		/// <summary>The previous state of the train doors</summary>
+		private DoorStates oldState;
+		/// <summary>Whether the pilot lamp is currently lit</summary>
+		internal bool Lit;
+
+		internal PilotLamp(AbstractCar car)
 		{
-			/// <summary>Played once when all doors are closed</summary>
-			internal CarSound OnSound;
-			/// <summary>Played once when the first door opens</summary>
-			internal CarSound OffSound;
-			/// <summary>Holds the reference to the base car</summary>
-			private readonly AbstractCar baseCar;
-			/// <summary>The previous state of the train doors</summary>
-			private DoorStates oldState;
-			/// <summary>Whether the pilot lamp is currently lit</summary>
-			internal bool Lit;
+			baseCar = car;
+			oldState = DoorStates.None;
+			OnSound = new CarSound();
+			OffSound = new CarSound();
+		}
 
-			internal PilotLamp(AbstractCar car)
+		internal void Update(DoorStates newState)
+		{
+			if (oldState != DoorStates.None & newState == DoorStates.None)
 			{
-				baseCar = car;
-				oldState = DoorStates.None;
-				OnSound = new CarSound();
-				OffSound = new CarSound();
+				Lit = true;
+				SoundBuffer buffer = OnSound.Buffer;
+				if (buffer != null)
+				{
+					Program.Sounds.PlaySound(buffer, 1.0, 1.0, OnSound.Position, baseCar, false);
+				}
 			}
-
-			internal void Update(DoorStates newState)
+			else if (oldState == DoorStates.None & newState != DoorStates.None)
 			{
-				if (oldState != DoorStates.None & newState == DoorStates.None)
+				Lit = false;
+				SoundBuffer buffer = OffSound.Buffer;
+				if (buffer != null)
 				{
-					Lit = true;
-					SoundBuffer buffer = OnSound.Buffer;
-					if (buffer != null)
-					{
-						Program.Sounds.PlaySound(buffer, 1.0, 1.0, OnSound.Position, baseCar, false);
-					}
+					Program.Sounds.PlaySound(buffer, 1.0, 1.0, OffSound.Position, baseCar, false);
 				}
-				else if (oldState == DoorStates.None & newState != DoorStates.None)
-				{
-					Lit = false;
-					SoundBuffer buffer = OffSound.Buffer;
-					if (buffer != null)
-					{
-						Program.Sounds.PlaySound(buffer, 1.0, 1.0, OffSound.Position, baseCar, false);
-					}
-				}
-				oldState = newState;
 			}
+			oldState = newState;
 		}
 	}
 }
